requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.28.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf
google-generativeai
langchain_experimental
langchain_google_genai
langchain_community
mysql-connector-python
pymysql
sentence-transformers
pandas